restructed Navigations and on clicking restaurant-info-card open new screen with restaurant info
