+++ package.json
@@ -13,6 +13,7 @@
     "@expo-google-fonts/oswald": "^0.2.0",
     "@react-navigation/bottom-tabs": "^6.0.5",
     "@react-navigation/native": "^6.0.2",
+    "@react-navigation/stack": "^6.0.7",
     "camelize": "^1.0.0",
     "expo": "~42.0.1",
     "expo-font": "~9.2.1",
@@ -20,6 +21,7 @@
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
+    "react-native-gesture-handler": "^1.10.3",
     "react-native-paper": "^4.9.2",
     "react-native-safe-area-context": "3.2.0",
     "react-native-screens": "~3.4.0",
