+++ src/features/restaurants/screens/restaurants-details.screen.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+import { RestaurantInfoCard } from "../components/restaurant-info-card.component";
+
+import { SafeArea } from "../../../components/utility/safe-area.component";
+
+export const RestaurantDetailScreen = ({ route }) => {
+  const { restaurant } = route.params;
+  return (
+    <SafeArea>
+      <RestaurantInfoCard restaurant={restaurant} />
+    </SafeArea>
+  );
+};
