+++ src/infrastructure/navigator/app.navigator.js
@@ -1,0 +1,66 @@
+import React from "react";
+import { NavigationContainer } from "@react-navigation/native";
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import { Text } from "react-native";
+import { Ionicons } from "@expo/vector-icons";
+
+import { RestaurantsNavigator } from "./restaurants.navigator";
+import { SafeArea } from "../../components/utility/safe-area.component";
+
+const Tab = createBottomTabNavigator();
+
+const Map = () => (
+  <SafeArea>
+    <Text>Map</Text>
+  </SafeArea>
+);
+
+const Settings = () => (
+  <SafeArea>
+    <Text>Settings</Text>
+  </SafeArea>
+);
+
+const TAB_ICON = {
+  Restaurants: {
+    Focused: "restaurant",
+    Unfocused: "restaurant-outline",
+  },
+  Map: {
+    Focused: "map",
+    Unfocused: "map-outline",
+  },
+  Settings: {
+    Focused: "settings",
+    Unfocused: "settings-outline",
+  },
+};
+
+const createScreenOptions = ({ route }) => {
+  return {
+    tabBarIcon: ({ focused, size, color }) => {
+      let iconName;
+
+      iconName = focused
+        ? TAB_ICON[route.name].Focused
+        : TAB_ICON[route.name].Unfocused;
+
+      return <Ionicons name={iconName} size={size} color={color} />;
+    },
+    tabBarActiveTintColor: "tomato",
+    tabBarInactiveColor: "grey",
+    headerShown: false,
+  };
+};
+
+export const AppNavigator = () => {
+  return (
+    <NavigationContainer>
+      <Tab.Navigator screenOptions={createScreenOptions}>
+        <Tab.Screen name="Restaurants" component={RestaurantsNavigator} />
+        <Tab.Screen name="Map" component={Map} />
+        <Tab.Screen name="Settings" component={Settings} />
+      </Tab.Navigator>
+    </NavigationContainer>
+  );
+};
