+++ src/infrastructure/navigator/restaurants.navigator.js
@@ -1,0 +1,27 @@
+import React from "react";
+import {
+  createStackNavigator,
+  TransitionPresets,
+} from "@react-navigation/stack";
+
+import { RestaurantScreen } from "../../features/restaurants/screens/restaurant.screen";
+import { RestaurantDetailScreen } from "../../features/restaurants/screens/restaurants-details.screen";
+
+const RestaurantStack = createStackNavigator();
+
+export const RestaurantsNavigator = () => {
+  return (
+    <RestaurantStack.Navigator
+      screenOptions={{
+        headerShown: false,
+        ...TransitionPresets.ModalPresentationIOS,
+      }}
+    >
+      <RestaurantStack.Screen name="Restaurant" component={RestaurantScreen} />
+      <RestaurantStack.Screen
+        name="RestaurantDetail"
+        component={RestaurantDetailScreen}
+      />
+    </RestaurantStack.Navigator>
+  );
+};
